+++ src/main/java/hs_burgenland/weather/controller/UserController.java
@@ -16,6 +16,9 @@ public class UserController {
 
     @PostMapping
     public ResponseEntity<?> createUser(@RequestBody final User user) {
+        if (user.getFirstname() == null || user.getLastname() == null) {
+            return ResponseEntity.badRequest().body("User must have a firstname and a lastname.");
+        }
         try {
             return ResponseEntity.ok(userService.createUser(user.getFirstname(), user.getLastname()));
         } catch (EntityAlreadyExistingException e) {
@@ -36,6 +39,10 @@ public class UserController {
 
     @GetMapping("/{id}")
     public ResponseEntity<?> getUserById(@PathVariable final int id) {
+        if (id < 0) {
+            return ResponseEntity.badRequest().body("User id must be at least 0.");
+        }
+
         try {
             return ResponseEntity.ok(userService.getUserById(id));
         } catch (EntityNotFoundException e) {
@@ -47,6 +54,11 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable final int id) {
+        if (id <= 0) {
+            // I want to prevent my default user from being deleted
+            return ResponseEntity.badRequest().body("User id must be greater than 0.");
+        }
+
         try {
             userService.deleteUser(id);
             return ResponseEntity.noContent().build();
